ownership / restricted deletes (#103)
* add "owner" user columns to gardens,entrypoints

* optional owner_identity_id field to create schemas

* assign default ownership in create functions

* only owners can delete gardens/entrypoints

* only draft DOIs are deletable

* factor out into assert_deletable util

* make owen default owner in migration script

* update deps; pull is_doi_registered into utils

* set doi draft status when creating garden

* fix: httpx uses a different kwarg than requests

* fix alembic migration bug

* fix migration script for devs who aren't owen

* make sure authed_user commits transaction

even if the route (e.g. /greet) does not

* error formatting

* poetry lock

* fix /garden tests

+++ garden-backend-service/src/api/routes/entrypoint.py
@@ -1,3 +1,5 @@
+from logging import getLogger
+
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -8,6 +10,9 @@ from src.api.schemas.entrypoint import (
     EntrypointMetadataResponse,
 )
 from src.models import Entrypoint, User
+from src.api.routes._utils import assert_deletable_by_user
+
+logger = getLogger(__name__)
 
 router = APIRouter(prefix="/entrypoint")
 
@@ -16,9 +21,26 @@ router = APIRouter(prefix="/entrypoint")
 async def add_entrypoint(
     entrypoint: EntrypointCreateRequest,
     db: AsyncSession = Depends(get_db_session),
-    _user: User = Depends(authed_user),
+    user: User = Depends(authed_user),
 ):
-    new_entrypoint = Entrypoint.from_dict(entrypoint.model_dump())
+    # default owner is authed_user unless owner_identity_id is explicitly provided
+    owner: User = user
+    if entrypoint.owner_identity_id is not None:
+        explicit_owner: User | None = await User.get(
+            db, identity_id=entrypoint.owner_identity_id
+        )
+        if explicit_owner is not None:
+            owner = explicit_owner
+        else:
+            logger.warning(
+                f"No user found with Globus identity ID {entrypoint.owner_identity_id}. "
+                f"Assigning default ownership to {user.identity_id} ({user.username}). "
+            )
+
+    new_entrypoint = Entrypoint.from_dict(
+        entrypoint.model_dump(exclude="owner_identity_id")
+    )
+    new_entrypoint.owner = owner
     db.add(new_entrypoint)
     try:
         await db.commit()
@@ -56,11 +78,12 @@ async def get_entrypoint_by_doi(
 async def delete_entrypoint(
     doi: str,
     db: AsyncSession = Depends(get_db_session),
-    _user: User = Depends(authed_user),
+    user: User = Depends(authed_user),
 ):
     entrypoint: Entrypoint | None = await Entrypoint.get(db, doi=doi)
 
     if entrypoint is not None:
+        assert_deletable_by_user(entrypoint, user)
         await db.delete(entrypoint)
         try:
             await db.commit()
