+++ garden-backend-service/src/api/routes/garden.py
@@ -1,21 +1,30 @@
+from logging import getLogger
+
 from fastapi import APIRouter, Depends, HTTPException, status
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy import select
+from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 from src.api.dependencies.auth import authed_user
 from src.api.dependencies.database import get_db_session
+from src.api.routes._utils import assert_deletable_by_user, is_doi_registered
 from src.api.schemas.garden import GardenCreateRequest, GardenMetadataResponse
-from src.models import Garden, User, Entrypoint
+from src.models import Entrypoint, Garden, User
 
 router = APIRouter(prefix="/garden")
+logger = getLogger(__name__)
 
 
 @router.post("", response_model=GardenMetadataResponse)
 async def add_garden(
     garden: GardenCreateRequest,
     db: AsyncSession = Depends(get_db_session),
-    _user: User = Depends(authed_user),
+    user: User = Depends(authed_user),
 ):
+    # if not specified, check draft status with the real world (doi.org)
+    if garden.doi_is_draft is None:
+        registered = await is_doi_registered(garden.doi)
+        garden.doi_is_draft = not registered
+
     # collect entrypoints by DOI
     stmt = select(Entrypoint).where(Entrypoint.doi.in_(garden.entrypoint_ids))
     result = await db.execute(stmt)
@@ -29,8 +38,26 @@ async def add_garden(
             status_code=404,
             detail=f"Failed to add garden. Could not find entrypoint(s) with DOIs: {missing_dois}",
         )
-    new_garden: Garden = Garden.from_dict(garden.model_dump(exclude="entrypoint_ids"))
+    # default owner is authed_user unless owner_identity_id is explicitly provided
+    owner: User = user
+    if garden.owner_identity_id is not None:
+        explicit_owner: User | None = await User.get(
+            db, identity_id=garden.owner_identity_id
+        )
+        if explicit_owner is not None:
+            owner = explicit_owner
+        else:
+            logger.warning(
+                f"No user found with Globus identity ID {garden.owner_identity_id}. "
+                f"Assigning default ownership to {user.identity_id} ({user.username}). "
+            )
+
+    new_garden: Garden = Garden.from_dict(
+        garden.model_dump(exclude={"entrypoint_ids", "owner_identity_id"})
+    )
+    new_garden.owner = owner
     new_garden.entrypoints = entrypoints
+
     db.add(new_garden)
     try:
         await db.commit()
@@ -65,10 +92,11 @@ async def get_garden_by_doi(
 async def delete_garden(
     doi: str,
     db: AsyncSession = Depends(get_db_session),
-    _user: User = Depends(authed_user),
+    user: User = Depends(authed_user),
 ):
     garden: Garden | None = await Garden.get(db, doi=doi)
     if garden is not None:
+        assert_deletable_by_user(garden, user)
         await db.delete(garden)
         try:
             await db.commit()
